% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katss.R
\name{kmer_presence}
\alias{kmer_presence}
\title{Count k-mer presence in a sequence}
\usage{
kmer_presence(
  input,
  kmer = 3,
  algo = c("regular", "shuffled"),
  bootstrap_iters = 0,
  sample = 25,
  seed = -1,
  klet = -1,
  sort = FALSE,
  threads = 1
)
}
\arguments{
\item{input}{Name of the file which you want to count k-mers presence from, or a
character vector/list containing sequences. Files must contain raw sequences,
FASTA, or FASTQ data. Gzip-compressed files are supported. A single character
string is interpreted as a file path; vectors with multiple elements and
lists are interpreted as raw sequences.}

\item{kmer}{Length of the k-mer you want to count. Currently, only k-mers up
to length 15 are supported.}

\item{algo}{Whether to perform regular counts, or count shuffled sequences.}

\item{bootstrap_iters}{Number of iterations to bootstrap.}

\item{sample}{Percent to subsample during bootstrap (should be between 0-100\%).}

\item{seed}{Specify the seed to be used by bootstrap. Since bootstrap
subsamples random sequences, seeding alters which random sequences will be
picked. This helps to ensure deterministic output which can be achieved by
using the same seed. To pick a random seed, set \code{seed=-1}.}

\item{klet}{Specify the k-let length to preserve during shuffling. This only
affects the output if \code{algo="shuffled"} is set. -1 chooses the default value.}

\item{sort}{Sort based on the counts from highest to lowest. By default,
the output given is sorted alphabetically on kmers (AA... first, TT... last).}

\item{threads}{Number of threads to use. Currently unsupported, changing the
value will make no difference in performance.}
}
\value{
Dataframe containing the counts for all k-mers.
}
\description{
Count the k-mer presence in a FASTQ, FASTA, raw sequences file, or a vector/list
of sequences. k-mer presence the number of sequences in which a k-mer appears.
That means that a k-mer is counted at most 1 time per sequnce (no repeats).
The final table then represents how many sequences a given k-mer is present in
given a dataset.
}
\examples{
# Load data
data(rbfox2_seqs)

# Count di-nucleotides from a vector of sequences
kmer_presence(rbfox2_seqs$bound, kmer = 2)

# Count a single sequence by wrapping it in a list, otherwise it will treat
# it as a filename
kmer_presence(list("ACGTACGT"), kmer = 2)

# Create temporary file with sequences
tf <- tempfile()
writeLines(rbfox2_seqs$bound, tf)

# Count presence of di-nucleotides in file
kmer_presence(tf, kmer = 2)

# Count presence of mono-nucleotides in file
kmer_presence(tf, kmer = 1)

# Count kmer presence of shuffled sequences
kmer_presence(tf, kmer = 1, algo = "shuffled")

# Specify k-let to preserve during shuffling
kmer_presence(tf, kmer = 1, algo = "shuffled", klet = 2)

# Count kmer presence with bootstrap
result <- kmer_presence(tf, bootstrap_iters = 100)
head(result)

# Subsample 55.55\% of the file per bootstrap iteration
result <- kmer_presence(tf, bootstrap_iters = 100, sample = 55.55)
head(result)

# Count presence of kmers in shuffled sequences with bootstrap
result <- kmer_presence(tf, algo = "shuffled", bootstrap_iters = 100)
head(result)

# Sort by number of counts, descending
result <- kmer_presence(tf, kmer = 5, sort = TRUE)
head(result)

# Cleanup file
unlink(tf)
}
