% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katss.R
\name{kmer_presence}
\alias{kmer_presence}
\title{Count k-mer presence in a sequence}
\usage{
kmer_presence(
  input,
  kmer = 3,
  algo = c("regular", "shuffled"),
  bootstrap_iters = 0,
  sample = 25,
  seed = -1,
  klet = -1,
  sort = FALSE,
  threads = 1
)
}
\arguments{
\item{input}{Name of the file which you want to count k-mers presence from, or a
character vector/list containing sequences. Files must contain raw sequences,
FASTA, or FASTQ data. Gzip-compressed files are supported. A single character
string is interpreted as a file path; vectors with multiple elements and
lists are interpreted as raw sequences.}

\item{kmer}{Length of the k-mer you want to count. Currently, only k-mers up
to length 15 are supported.}

\item{algo}{Whether to perform regular counts, or count shuffled sequences.}

\item{bootstrap_iters}{Number of iterations to bootstrap.}

\item{sample}{Percent to subsample during bootstrap (should be between 0-100\%).}

\item{seed}{Specify the seed to be used by bootstrap. Since bootstrap
subsamples random sequences, seeding alters which random sequences will be
picked. This helps to ensure deterministic output which can be achieved by
using the same seed. To pick a random seed, set \code{seed=-1}.}

\item{klet}{Specify the k-let length to preserve during shuffling. This only
affects the output if \code{algo="shuffled"} is set. -1 chooses the default value.}

\item{sort}{Sort based on the counts from highest to lowest. By default,
the output given is sorted alphabetically on kmers (AA... first, TT... last).}

\item{threads}{Number of threads to use. Currently unsupported, changing the
value will make no difference in performance.}
}
\value{
Dataframe containing the counts for all k-mers.
}
\description{
Count the k-mer presence in a FASTQ, FASTA, raw sequences file, or a vector/list
of sequences. k-mer presence the number of sequences in which a k-mer appears.
That means that a k-mer is counted at most 1 time per sequnce (no repeats).
The final table then represents how many sequences a given k-mer is present in
given a dataset.
}
